feat: Add Docker support and backend improvements
This commit introduces Docker support for the backend API, enabling easier local development and deployment. It also includes several backend improvements such as:

- **Dockerization:** A `Dockerfile` and `docker-compose.yml` are added for containerizing the backend and MongoDB.
- **Environment Variables:** A `.env.example` file is created for managing environment variables.
- **Code Refactoring:** The backend server logic is refactored into `app.js` for better organization.
- **Testing:** A basic health check test is added using Jest.
- **Dependency Updates:** Minor updates to dependencies like `multer`.
- **Documentation:** The README is updated with instructions for running the backend with Docker.
