+++ backend/.dockerignore
@@ -1,0 +1,9 @@
+node_modules
+npm-debug.log
+.env
+uploads
+coverage
+__tests__
+.git
+.DS_Store
+
