+++ backend/.env.example
@@ -1,3 +1,11 @@
+NODE_ENV=development
+PORT=3001
+MONGODB_URI=mongodb://localhost:27017/connectfe
+JWT_SECRET=change_me
+EXPO_ACCESS_TOKEN=
+EXPO_PROJECT_ID=
+CLOUDINARY_URL=
+
 # ===== CONNECTFÉ - VARIÁVEIS DE AMBIENTE =====
 
 # Configurações do Servidor
