+++ backend/package.json
@@ -22,7 +22,7 @@
     "igrejas-app-design": "file:..",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.0.3",
-    "multer": "^1.4.7",
+    "multer": "1.4.5-lts.1",
     "natural": "^6.12.0",
     "node-cron": "^3.0.3",
     "socket.io": "^4.8.1",
